OPL-7: Invoice refactor. Added base classes for services and validation
